feat: Refactor project structure and implement core functionalities
- Added configuration management with environment loading and LangSmith setup.
- Introduced model management with available models and provider definitions.
- Implemented document processing for PDF files, including metadata handling and chunking.
- Developed LLM management to interface with various providers (OpenAI, Claude, Upstage, Ollama).
- Created QA chain for conversational retrieval and response generation.
- Established vector store management for document embeddings and retrieval.
- Added utility functions for processing PDFs in folders and formatting documents with citations.
- Implemented company comparison tool for financial metrics.
- Removed deprecated pipeline configuration and QA pipeline files.
