+++ config/__init__.py
@@ -1,0 +1,3 @@
+# config/__init__.py
+from .env import load_environment, configure_langsmith, LANGSMITH_PROJECT
+from .models import AVAILABLE_MODELS, AVAILABLE_MODELS_EMBEDDINGS, check_model_name
