+++ core/__init__.py
@@ -1,0 +1,6 @@
+# core/__init__.py
+__all__ = ["LLMManager", "DocumentProcessor", "VectorStoreManager", "QAChain", "QAMemoryChain"]
+from .llm_manager import LLMManager
+from .document_processor import DocumentProcessor
+from .vector_store import VectorStoreManager
+from .qa_chain import QAChain, QAMemoryChain
