+++ tools/__init__.py
@@ -1,0 +1,1 @@
+from .company_compare import compare_companies
